requirements.txt
torch
opencv-python
numpy
joblib
streamlit
Pillow
transformers
ultralytics
scikit-learn